Set fingerprint also in the client
This is needed because otherwise the call to _wait_for_handshake() inside
the method WiFiClientSecure::_connectSSL() will fail with the error:
BSSL:Couldn't connect. Error = 'Chain could not be linked to a trust anchor.'

This is the call chain:
BlynkArduinoClientGen<Client>::connect()
 WiFiClientSecure::connect()
  WiFiClientSecure::_connectSSL()

BSSL:Couldn't connect. Error = 'Chain could not be linked to a trust anchor.'